github_dw/requirements.txt
psycopg2-binary
requests
streamlit
groq
pandas
python-dotenv
